requirements.txt
google-generativeai==0.8.5
python-dotenv==1.1.1
google-cloud-aiplatform
